requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
pyJWT